code and folder restructure
